feat(settings): reorganize into 5 tabs with extracted render functions
Restructure settings from 4 tabs to 5 intent-based tabs:
- Setup (auth, save location)
- Content (what to pull from Reddit)
- Output (how files look — organization, comments, templates, media, Obsidian integration)
- Filters (unchanged)
- Advanced (performance, link preservation, progress tracking)

Key improvements:
- Unify comment settings (previously split across Import and Advanced)
- Extract each tab into its own render function (settings.ts: 1337 → 96 lines)
- Use createCollapsibleSection() for all collapsible content
- Improve setting descriptions with practical guidance
- Migrate 'import' tab selection to 'content' on load
